AdminController as must be

+++ NB-API/Controllers/AdminController.cs
@@ -17,29 +17,31 @@ namespace NB_API.Controllers
         }
 
        // [HttpDelete("Oprydning/Brugere")]
-        [HttpGet("Oprydning/Brugere/{id}")]
+        [HttpDelete("Oprydning/Brugere/{id}")]
         public async Task<IActionResult> CleanupDeletedUsers(int id)
         {
-            int deleted = 0;
-            var result = _context.Bruger.Where(b => b.DeleteTime.Value.AddDays(92) < DateTime.Now.Date ).ToList();
+            var result =  _context.Bruger.Where(b => b.DeleteTime != null && b.DeleteTime.Value.AddDays(92) < DateTime.Now.Date ).ToList();
             if (result == null)
             {
                 return NoContent();
             }
             try
             {
-                deleted = result.Count();
-                foreach(var item in result)
+                if (result != null)
                 {
-                    _context.Bruger.Remove(item);
+                    foreach(var item in result)
+                    {
+                       _context.Bruger.Remove(item);
+                    }
+                    await _context.SaveChangesAsync();
+
                 }
-                await _context.SaveChangesAsync();
-                return Ok(deleted.ToString());
+                return NoContent();
             }
             catch (Exception e)
             {
 
-                return BadRequest("Dette gik galt: " + e.Message);
+                return BadRequest("Dette gik galt: " + e.InnerException.Message);
             }
         }
     }
